+++ frontend/src/components/layouts/Footer.jsx
@@ -1,0 +1,6 @@
+
+export default function Footer() {
+    return (
+        <footer className={"h-5"}></footer>
+    );
+}
